[Refactor](checkComputedOptionFactory): Update import path for getPathManager
