+++ packages/canard/schema-form/src/core/nodes/AbstractNode/utils/computeFactory/utils/__tests__/errorHandling.test.ts
@@ -6,7 +6,7 @@ import type { JsonSchemaWithVirtual } from '@/schema-form/types';
 import { checkComputedOptionFactory } from '../checkComputedOptionFactory';
 import { getConditionIndexFactory } from '../getConditionIndexFactory';
 import { getObservedValuesFactory } from '../getObservedValuesFactory';
-import { getPathManager } from '../getPathManager/getPathManager';
+import { getPathManager } from '../getPathManager';
 
 describe('Error Handling in Dynamic Function Creation', () => {
   describe('checkComputedOptionFactory', () => {
