+++ packages/canard/schema-form/src/core/nodes/AbstractNode/utils/computeFactory/utils/__tests__/getConditionIndexFactory.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from 'vitest';
 import type { JsonSchemaWithVirtual } from '@/schema-form/types';
 
 import { getConditionIndexFactory } from '../getConditionIndexFactory';
-import { getPathManager } from '../getPathManager/getPathManager';
+import { getPathManager } from '../getPathManager';
 
 describe('getConditionIndexFactory', () => {
   it('유효하지 않은 스키마에 대해 undefined를 반환해야 함', () => {
